Move version info away from Designer
 - This allows consumers (ie: pmd-cli) to access the Designer version without requiring JavaFX
